+++ src/perft.h
@@ -3,15 +3,17 @@
  *
  * @brief Move generator test header file.
  *
- * @date 1998 - 2017
+ * @date 1998 - 2024
  * @author Richard Delorme
- * @version 4.4
+ * @author Toshihiko Okuhara
+ * @version 4.6
  */
 
 #ifndef EDAX_PERFT_H
 #define EDAX_PERFT_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 struct Board;
 struct Line;
@@ -19,8 +21,8 @@ void count_games(const struct Board*, const int);
 void quick_count_games(const struct Board*, const int, const int);
 void count_positions(const struct Board*, const int, const int);
 void count_shapes(const struct Board*, const int, const int);
-void estimate_games(const struct Board*, const long long);
-void seek_highest_mobility(const struct Board*, const unsigned long long);
+void estimate_games(const struct Board*, const int64_t);
+void seek_highest_mobility(const struct Board*, const uint64_t);
 bool seek_position(const struct Board*, const struct Board*, struct Line*);
 
 /** HashTable of positions */
